ENHANCEMENT: Only hide children from SiteTree in CMSMain
Applying a universal exclusion filter has the unintended consequence of filtering out TreeDropdownFields (e.g. in the EditToolbar link form in HtmlEditorField), meaning these pages are invisible and impossible to link to. This is a bit of an ugly fix, but it seems like the exclusion of the children should be on a white list of controllers rather than a global mutation.